+++ R/anno.R
@@ -30,12 +30,15 @@ getalign <- function(mz1,mz2,rt1 = NULL,rt2 = NULL,ppm=10,deltart=10){
                 overlaprt <- data.table::foverlaps(rta, rtb, which = TRUE)
                 overlaprt$rt1 <- rt1[overlaprt$xid]
                 overlaprt$rt2 <- rt2[order(rt2,decreasing = F)][overlaprt$yid]
-                over <- merge(overlapms,overlaprt,by = 'xid')
+                over <- merge(overlapms,overlaprt,by = 'xid',allow.cartesian=TRUE)
                 over2 <- over[abs(over$rt1-over$rt2)<deltart,]
                 over3 <- over2[,-c(2,5)]
                 re <- data.frame(over3[!duplicated(over3)&stats::complete.cases(over3)])
+                mr <- paste(mz2,rt2)
+                mr2 <- paste(re$mz2,re$rt2)
+
                 if(nrow(re)>0){
-                        return(re)
+                        return(re[mr2 %in% mr,])
                 }else{
                         message('No result could be found.')
                 }
@@ -75,40 +78,7 @@ getoverlappeak <- function(list1, list2) {
         index <- (!is.na(overlapms)) & (!is.na(overlaprt))
         return(index)
 }
-#' Get the overlap peaks by mass range
-#' @param mzrange1 mass range 1 to be overlapped
-#' @param mzrange2 mass range 2 to overlap
-#' @return logical index for mzrange1's peaks
-#' @export
-#' @seealso \code{\link{getmzrt}}, \code{\link{getimputation}}, \code{\link{getmr}},\code{\link{getdoe}}, \code{\link{getoverlappeak}},\code{\link{getoverlaprt}}
-getoverlapmass <- function(mzrange1, mzrange2) {
-        mz1 <- data.table::as.data.table(mzrange1)
-        mz2 <- data.table::as.data.table(mzrange2)
-        colnames(mz1) <- colnames(mz2) <- c('min', 'max')
-        data.table::setkey(mz2, min, max)
-        overlapms <-
-                data.table::foverlaps(mz1, mz2, which = TRUE, mult = 'first')
 
-        index <- (!is.na(overlapms))
-        return(index)
-}
-#' Get the overlap peaks by retention time
-#' @param rtrange1 mass range 1 to be overlapped
-#' @param rtrange2 mass range 2 to overlap
-#' @return logical index for rtrange1's peaks
-#' @export
-#' @seealso \code{\link{getmzrt}}, \code{\link{getimputation}}, \code{\link{getmr}},\code{\link{getdoe}}, \code{\link{getoverlapmass}},\code{\link{getoverlappeak}}
-getoverlaprt <- function(rtrange1, rtrange2) {
-        rt1 <- data.table::as.data.table(rtrange1)
-        rt2 <- data.table::as.data.table(rtrange2)
-        colnames(rt1) <- colnames(rt2) <- c('min', 'max')
-        data.table::setkey(rt2, min, max)
-        overlapms <-
-                data.table::foverlaps(rt1, rt2, which = TRUE, mult = 'first')
-
-        index <- (!is.na(overlapms))
-        return(index)
-}
 #' Annotation of MS1 data by compounds database by predefined paired mass distance
 #' @param pmd adducts formula or paired mass distance for ions
 #' @param mz unknown mass to charge ratios vector