+++ NEWS.md
@@ -4,7 +4,7 @@
 - remove the export of deprecated function
 - fix the typo in documents
 - improve getMSP to fit MoNA MSP format
-- add getalign function to align two peaks vectors
+- add getalign function to align two peaks vectors and remove getoverlapmass, getoverlaprt
 - modify getmass to accept negative value
 - add getms1anno for MS1 annotation by local compounds database with predefined paired mass distances
 - add support to save target data csv without change default name